+++ Backend/src/features/printing-editions/services/printingEditionService.ts
@@ -3,44 +3,54 @@ import * as repository  from "../repositories/printingEditionRepository";
 import { PrintingEditionFilterModel } from "../../shared/filterModels/printingEditionFilterModel";
 import { BaseResponse } from "../../shared/db-models/BaseResponse";
 import { PrintingEdition } from "../api";
+import printingEditionValidateSchema from "../operations/PrintingEditionRequest.schema.json";
+import { validateWithJsonSchema } from "../../utils/validateWithJsonSchema";
+import idValidateSchema from "../../utils/IdRequest.schema.json"
 
-let printingEdition = new printingEditionModel();
- printingEdition.errors = new Array<string>();
 
-export async function createAsync(printingEditionParam: printingEditionModel): Promise<Array<string>> {
-
-    if (printingEditionParam == null) {
-        printingEdition.errors.push('fields are empty');
-        return printingEdition.errors;
+export async function createAsync(printingEditionParam: printingEditionModel): Promise<any> {
+     const validateResult = validateWithJsonSchema(printingEditionValidateSchema,validateWithJsonSchema)
+    if (!validateResult.valid) {
+        return {message: "printingEdition parameters is not valide" , error: validateResult.errors}
     }
 
     const result = await repository.createAsync(printingEditionParam);
     
     if (!result) {
-        printingEdition.errors.push("failed to save document");
+       return ("failed to save document");
     }
 
-    return printingEdition.errors;
+    return result;
 }
 
-export async function removeAsync(id: string): Promise<Array<string>> {
+export async function removeAsync(id: string): Promise<any> {
+    const validateResult = validateWithJsonSchema(id,idValidateSchema);
+    
+    if (!validateResult.valid) {
+        return {message: "id parameters is not valide" , error: validateResult.errors}
+    }
+
     const result = await repository.removeAsync(id);
     
     if (!result) {
-        printingEdition.errors.push('product did not find');
+      return result    
     }
 
-    return printingEdition.errors;
+    return result;
 }
 
-export async function updateAsync(printingEditionParam: printingEditionModel): Promise<Array<string>> {
+export async function updateAsync(printingEditionParam: printingEditionModel): Promise<any> {
+    const validateResult = validateWithJsonSchema(printingEditionValidateSchema,validateWithJsonSchema)
+    if (!validateResult.valid) {
+        return {message: "printingEdition parameters is not valide" , error: validateResult.errors}
+    }
     const result = await repository.updateAsync(printingEditionParam);
     
     if (!result) {
-        printingEdition.errors.push('product did not find');
+        return "product did not update";
     }
 
-    return printingEdition.errors;
+    return result;
 }
 
 export async function getAsync(filter: PrintingEditionFilterModel): Promise<BaseResponse<PrintingEdition>> {
